+++ src/main/java/de/samply/resolvers/RequestBodyCache.java
@@ -1,5 +1,6 @@
 package de.samply.resolvers;
 
+import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.context.annotation.Scope;
@@ -8,18 +9,27 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.context.WebApplicationContext;
 
 import java.io.IOException;
+import java.util.Collections;
 import java.util.Map;
 
 @Component
 @Scope(value = WebApplicationContext.SCOPE_REQUEST, proxyMode = ScopedProxyMode.TARGET_CLASS)
 public class RequestBodyCache {
 
+    private final static ObjectMapper objectMapper = new ObjectMapper();
     private Map<String, Object> jsonBody;
 
     public Map<String, Object> getJsonBody(HttpServletRequest request) throws IOException {
         if (jsonBody == null) {
-            // Parse the JSON body only once per request
-            jsonBody = new ObjectMapper().readValue(request.getInputStream(), Map.class);
+            try {
+                jsonBody = objectMapper.readValue(
+                        request.getInputStream(),
+                        new TypeReference<>() {
+                        }
+                );
+            } catch (com.fasterxml.jackson.databind.exc.MismatchedInputException e) {
+                jsonBody = Collections.emptyMap();
+            }
         }
         return jsonBody;
     }